feat: implement module task
